# servers/ai-gateway/contracts/gen-aigw.openapi.yaml
openapi: 3.1.0

# ============================================================================
# 契约③ generation-service ↔ ai-gateway（内部契约）
# 提供方 = servers/ai-gateway（消费方 = servers/generation-service）
# 范围 = F1 首刀 A1，仅首刀字段；与 feature 总 spec §2 主链 / §5.1 / §6.4 一致。
# 关键设定（spec §6.4）：ai-gateway = 薄代理 + 横切。本契约把「系统指令」与
#   「性格(数据槽)」作为**分立字段**传输，保持「指令 vs 用户数据」物理分离；
#   gateway 仅做协议适配 + 通用横切 guardrail（输出审核第三层），**不解析业务字段**。
# A1 不含：多家 provider / fallback、预算上限、精细限流。
# ============================================================================

info:
  title: generation-service ↔ ai-gateway 契约（A1 首刀）
  version: 0.1.1-a1
  description: |
    servers/generation-service（消费方）经 servers/ai-gateway（提供方，唯一 LLM 入口）
    调用单家 LLM（DeepSeek）做补全。gateway 持有 LLM key、协议适配、超时/重试，
    并对输出施加通用横切 guardrail（注入回显 / 泄漏 prompt / 不安全内容过滤）后返回。

    错误码口径（F4·GW-1 评审 ③ 后演化，向后兼容·加 400/500）：所有非 2xx 一律
    用 Error{code,message} 体，机读 code 区分语义——
      400 invalid_request    请求不符 schema（缺字段 / 超长 / 多余字段）；
      422 guardrail_blocked  **仅** 输出横切 guardrail 拦截（不再混入请求校验）；
      500 internal_error     网关内部异常（兜底，绝不回显 prompt/personality）；
      502 upstream_unavailable 上游 DeepSeek 不可用 / 超时（重试后仍失败）。
    最坏时延：单次上游超时 15s + 最多 1 次重试（退避 0.5s）⇒ 网关最坏 ≈30s；
    消费方（gen）应据此设自身超时 > 30s、且不再叠加重试（G-1 非幂等，避免重复烧配额）。

servers:
  - url: /
    description: servers/ai-gateway（默认 :8001）

paths:
  /v1/complete:
    post:
      operationId: complete
      summary: LLM 补全（输出经横切 guardrail）
      description: |
        对应主链「gen→ai-gateway（内部：LLM 调用）」。
        gateway 把 systemInstruction 映射为系统指令、personality 作为数据槽内容、
        realTime 作为现实时间上下文，调 DeepSeek；对返回文本施加通用横切 guardrail
        后返回。guardrail 命中不安全/注入回显时以 422 拒绝。
        F4·GW-1（守"薄"硬约束）：① guardrail 为 **block-only**——命中即 422，
        **不做静默改写/脱敏**（改写=篡改业务产出）；② 上述字段→角色为**固定机械映射**，
        gateway **不按内容做分支判断**（详见 spec §6.4）。
        错误码：请求不符 schema → 400 invalid_request（与 guardrail 的 422 分离，
        评审 ③·C1）；guardrail 命中 → 422 guardrail_blocked；内部异常 → 500
        internal_error（兜底脱敏，评审 ③·C3）；上游失败 → 502 upstream_unavailable。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompletionRequest'
      responses:
        '200':
          description: 补全文本（已过输出横切 guardrail）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompletionResponse'
        '400':
          description: 请求不符 schema（缺字段 / 超长 / 多余字段）；code=invalid_request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: 输出横切 guardrail 拦截（注入回显 / 泄漏 prompt / 不安全内容）；code=guardrail_blocked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 网关内部异常兜底（绝不回显 prompt/personality）；code=internal_error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '502':
          description: 上游 LLM（DeepSeek）不可用 / 超时（重试后仍失败）；code=upstream_unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:

    CompletionRequest:
      type: object
      additionalProperties: false
      required: [systemInstruction, personality, realTime]
      properties:
        systemInstruction:
          type: string
          minLength: 1
          description: |
            系统指令（由 gen 构建）。指令文本，**不含用户数据**——
            用户性格不拼进此字段，单独走 personality 数据槽。
        personality:
          type: string
          maxLength: 4000
          description: |
            性格（数据槽）。作为**数据**传输，与 systemInstruction 物理分离；
            gateway 将其置入数据/用户消息，不并入指令。maxLength 仅为稳定性横切
            （防超大 payload），非业务录入审核（录入审核在 api/gen）。
        realTime:
          type: string
          format: date-time
          description: 现实时间锚点（设定城市时区），作为生成上下文。

    CompletionResponse:
      type: object
      additionalProperties: false
      required: [text]
      properties:
        text:
          type: string
          description: LLM 生成文本，已通过输出横切 guardrail。

    Error:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
          description: 机读错误码（如 guardrail_blocked / upstream_unavailable）。
        message:
          type: string